Remove .dev and .prod configs
