+++ packages/webpack-config/src/index.ts
@@ -1,23 +1,17 @@
 import { Configuration } from 'webpack';
-import { Environment, InputEnvironment, Arguments, DevConfiguration } from './types';
-import developmentConfig from './webpack.dev';
-import productionConfig from './webpack.prod';
-import { getMode } from './utils';
+
+import { Arguments, DevConfiguration, Environment, InputEnvironment } from './types';
 import * as Diagnosis from './utils/Diagnosis';
 import { validateEnvironment } from './utils/validate';
+import webpackConfig from './webpack.config';
 
 export default async function(
   env: InputEnvironment,
   argv: Arguments = {}
 ): Promise<Configuration | DevConfiguration> {
   const environment: Environment = validateEnvironment(env);
-  const mode = getMode(environment);
-  let config: Configuration | DevConfiguration;
-  if (mode === 'development') {
-    config = await developmentConfig(environment, argv);
-  } else {
-    config = await productionConfig(environment, argv);
-  }
+
+  const config: Configuration | DevConfiguration = await webpackConfig(environment, argv);
 
   if (environment.info) {
     Diagnosis.reportAsync(config, environment);
